+++ src/core/dataloader/dataloader.interface.ts
@@ -2,7 +2,7 @@ import DataLoader from 'dataloader';
 import { ILocation } from '@domain/common/location/location.interface';
 import { IReference } from '@domain/common/reference/reference.interface';
 import { ITagset } from '@domain/common/tagset/tagset.interface';
-import { IProfile } from '@domain/community/profile/profile.interface';
+import { IProfile } from '@domain/common/profile/profile.interface';
 import { ICallout } from '@domain/collaboration/callout/callout.interface';
 import { IRelation } from '@domain/collaboration/relation/relation.interface';
