+++ src/core/dataloader/dataloader.module.ts
@@ -1,5 +1,5 @@
 import { OrganizationModule } from '@domain/community/organization/organization.module';
-import { ProfileModule } from '@domain/community/profile/profile.module';
+import { ProfileModule } from '@domain/common/profile/profile.module';
 import { UserModule } from '@domain/community/user/user.module';
 import { CollaborationModule } from '@domain/collaboration/collaboration/collaboration.module';
 import { Module } from '@nestjs/common';
