+++ src/core/validation/handlers/base/base.handler.ts
@@ -41,11 +41,11 @@ import { CommunicationCreateDiscussionInput } from '@domain/communication/commun
 import { CreateFeedbackOnCommunityContextInput } from '@domain/community/community/dto/community.dto.create.feedback.on.context';
 import { CreateReferenceOnCardProfileInput } from '@domain/collaboration/card-profile/dto/card.profile.dto.create.reference';
 import { CreateReferenceOnContextInput } from '@domain/context/context/dto/context.dto.create.reference';
-import { CreateReferenceOnProfileInput } from '@domain/community/profile/dto/profile.dto.create.reference';
+import { CreateReferenceOnProfileInput } from '@domain/common/profile/dto/profile.dto.create.reference';
 import {
   CreateTagsetOnProfileInput,
   UpdateProfileInput,
-} from '@domain/community/profile/dto';
+} from '@domain/common/profile/dto';
 import { ApplicationEventInput } from '@domain/community/application/dto/application.dto.event';
 import { CanvasCheckoutEventInput } from '@domain/common/canvas-checkout/dto/canvas.checkout.dto.event';
 import { OrganizationVerificationEventInput } from '@domain/community/organization-verification/dto/organization.verification.dto.event';
