+++ src/domain/common/profile/profile.interface.ts
