+++ src/domain/common/profile/profile.service.ts
@@ -13,8 +13,8 @@ import { IReference } from '@domain/common/reference/reference.interface';
 import { ReferenceService } from '@domain/common/reference/reference.service';
 import { ITagset } from '@domain/common/tagset/tagset.interface';
 import { TagsetService } from '@domain/common/tagset/tagset.service';
-import { Profile } from '@domain/community/profile/profile.entity';
-import { IProfile } from '@domain/community/profile/profile.interface';
+import { Profile } from '@domain/common/profile/profile.entity';
+import { IProfile } from '@domain/common/profile/profile.interface';
 import { AuthorizationPolicy } from '@domain/common/authorization-policy';
 import { AuthorizationPolicyService } from '@domain/common/authorization-policy/authorization.policy.service';
 import { VisualService } from '@domain/common/visual/visual.service';
