+++ src/domain/common/reference/reference.entity.ts
@@ -1,7 +1,7 @@
 import { Column, Entity, ManyToOne } from 'typeorm';
 import { IReference } from './reference.interface';
 import { Context } from '@domain/context/context/context.entity';
-import { Profile } from '@domain/community/profile/profile.entity';
+import { Profile } from '@domain/common/profile/profile.entity';
 import { AuthorizableEntity } from '@domain/common/entity/authorizable-entity';
 import { CardProfile } from '@domain/collaboration/card-profile/card.profile.entity';
