+++ src/domain/common/visual/visual.entity.ts
@@ -2,7 +2,7 @@ import { Column, Entity, ManyToOne } from 'typeorm';
 import { Context } from '@domain/context/context/context.entity';
 import { AuthorizableEntity } from '@domain/common/entity/authorizable-entity';
 import { IVisual } from './visual.interface';
-import { Profile } from '@domain/community/profile/profile.entity';
+import { Profile } from '@domain/common/profile/profile.entity';
 
 @Entity()
 export class Visual extends AuthorizableEntity implements IVisual {
