+++ src/domain/community/contributor/contributor.entity.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-inferrable-types */
 import { JoinColumn, OneToOne } from 'typeorm';
-import { Profile } from '@domain/community/profile/profile.entity';
+import { Profile } from '@domain/common/profile/profile.entity';
 import { Agent } from '@domain/agent/agent/agent.entity';
 import { IContributor } from './contributor.interface';
 import { NameableEntity } from '@domain/common/entity/nameable-entity/nameable.entity';
