+++ src/domain/community/contributor/contributor.interface.ts
@@ -1,4 +1,4 @@
-import { IProfile } from '@domain/community/profile/profile.interface';
+import { IProfile } from '@domain/common/profile/profile.interface';
 import { IAgent } from '@domain/agent/agent/agent.interface';
 import { ObjectType } from '@nestjs/graphql';
 import { INameable } from '@domain/common/entity/nameable-entity/nameable.interface';
