+++ src/domain/community/contributor/dto/contributor.dto.create.ts
@@ -1,5 +1,5 @@
 import { Field, InputType } from '@nestjs/graphql';
-import { CreateProfileInput } from '@domain/community/profile/dto';
+import { CreateProfileInput } from '@domain/common/profile/dto';
 import { ValidateNested } from 'class-validator';
 import { Type } from 'class-transformer';
 import { CreateNameableInput } from '@domain/common/entity/nameable-entity/dto/nameable.dto.create';
