+++ src/domain/community/contributor/dto/contributor.dto.update.ts
@@ -1,5 +1,5 @@
 import { Field, InputType } from '@nestjs/graphql';
-import { UpdateProfileInput } from '@domain/community/profile/dto/profile.dto.update';
+import { UpdateProfileInput } from '@domain/common/profile/dto/profile.dto.update';
 import { IsOptional, ValidateNested } from 'class-validator';
 import { Type } from 'class-transformer';
 import { UpdateNameableInput } from '@domain/common/entity/nameable-entity/dto/nameable.dto.update';
