+++ src/domain/community/index.ts
@@ -1,5 +1,5 @@
 export * from './application';
 export * from './organization';
-export * from './profile';
+export * from '../common/profile';
 export * from './user';
 export * from './user-group';
