+++ src/domain/community/user-group/user-group.entity.ts
@@ -1,7 +1,7 @@
 import { Organization } from '@domain/community/organization/organization.entity';
 import { Column, Entity, JoinColumn, ManyToOne, OneToOne } from 'typeorm';
 import { IUserGroup } from '@domain/community/user-group/user-group.interface';
-import { Profile } from '@domain/community/profile/profile.entity';
+import { Profile } from '@domain/common/profile/profile.entity';
 import { Community } from '@domain/community/community/community.entity';
 import { AuthorizableEntity } from '@domain/common/entity/authorizable-entity';
