+++ src/domain/community/user-group/user-group.interface.ts
@@ -1,5 +1,5 @@
 import { IAuthorizable } from '@domain/common/entity/authorizable-entity';
-import { IProfile } from '@domain/community/profile/profile.interface';
+import { IProfile } from '@domain/common/profile/profile.interface';
 import { Field, ObjectType } from '@nestjs/graphql';
 @ObjectType('UserGroup')
 export abstract class IUserGroup extends IAuthorizable {
