+++ src/domain/community/user/user.service.authorization.ts
@@ -7,7 +7,7 @@ import {
 import { IUser } from '@domain/community/user';
 import { AgentService } from '@domain/agent/agent/agent.service';
 import { UserService } from './user.service';
-import { ProfileAuthorizationService } from '@domain/community/profile/profile.service.authorization';
+import { ProfileAuthorizationService } from '@domain/common/profile/profile.service.authorization';
 import { IAuthorizationPolicy } from '@domain/common/authorization-policy';
 import { EntityNotInitializedException } from '@common/exceptions';
 import { AuthorizationPolicyService } from '@domain/common/authorization-policy/authorization.policy.service';
